+++ starter/app/src/main/java/com/example/android/politicalpreparedness/representative/model/VoterInfo.kt
@@ -1,4 +1,4 @@
-package com.example.android.politicalpreparedness.network.models
+package com.example.android.politicalpreparedness.representative.model
 
 class VoterInfo (
     val correspondenceStateName: String,
